chore: fix failing testpypi workflow
